Hide the folder itself and its children from the move dialog
This commit ensures that when a user tries to move a folder, the folder itself and its subfolders are not listed as possible destinations in the "Move to" dialog.

This is achieved by modifying the `getFoldersWithParentsExcept` function in `HomeViewModel.kt` to filter out the selected folder and its children when generating the list of available folders. The `LeftSideMenu.kt` is updated to use this new function.